Move path_get to separate module
